Fix FS#13009.
This file revealed several problems with our ASF parser:

1)  The packet count in the ASF was actually a 64 bit value,
leading to overflow in very long files.

2)  Seeking blindly trusted the bitrate listed in the ASF header
rather than computing it from the packet size and number of packets.

Fix these problems and fix a few minor issues.

+++ lib/rbcodec/codecs/libasf/asf.h
@@ -33,7 +33,7 @@ struct asf_waveformatex_s {
     uint16_t blockalign;
     uint16_t bitspersample;
     uint16_t datalen;
-    uint16_t numpackets;
+    uint64_t numpackets;
     uint8_t data[46];
 };
 typedef struct asf_waveformatex_s asf_waveformatex_t;
